feature/security рефакторинг
