+++ src/main/java/ru/froleod/FinanceTrackerApp/controller/FinancialGoalPageController.java
@@ -9,19 +9,14 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import ru.froleod.FinanceTrackerApp.model.BankAccount;
 import ru.froleod.FinanceTrackerApp.model.FinancialGoal;
 import ru.froleod.FinanceTrackerApp.model.Transaction;
-import ru.froleod.FinanceTrackerApp.model.enums.GoalStatus;
 import ru.froleod.FinanceTrackerApp.model.enums.TransactionType;
-import ru.froleod.FinanceTrackerApp.repo.FinancialGoalRepository;
-import ru.froleod.FinanceTrackerApp.service.BankAccountService;
 import ru.froleod.FinanceTrackerApp.service.FinancialGoalService;
 import ru.froleod.FinanceTrackerApp.service.TransactionService;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.time.LocalDate;
 import java.util.List;
 
 @Controller
@@ -32,25 +27,20 @@ public class FinancialGoalPageController {
     private FinancialGoalService financialGoalService;
 
     @Autowired
-    private FinancialGoalRepository financialGoalRepository;
-
-    @Autowired
     private TransactionService transactionService;
 
-    @Autowired
-    private BankAccountService bankAccountService;
 
     @GetMapping
     public String viewFinancialGoals(Model model) {
         List<FinancialGoal> financialGoals = financialGoalService.getAllFinancialGoals();
         model.addAttribute("financialGoals", financialGoals);
-        return "financial-goals";
+        return "financialGoals/financial-goals";
     }
 
     @GetMapping("/create")
     public String createFinancialGoalForm(Model model) {
         model.addAttribute("financialGoal", new FinancialGoal());
-        return "create-financial-goal";
+        return "financialGoals/create-financial-goal";
     }
 
     @PostMapping
@@ -61,27 +51,8 @@ public class FinancialGoalPageController {
             @RequestParam String startDate,
             @RequestParam String endDate,
             @RequestParam String status) {
-
-        FinancialGoal financialGoal = new FinancialGoal();
-
-        financialGoal.setName(name);
-        financialGoal.setTargetAmount(targetAmount);
-
-        BankAccount bankAccount = bankAccountService.getBankAccountById(bankAccountId)
-                .orElseThrow(() -> new RuntimeException("BankAccount not found"));
-        financialGoal.setBankAccount(bankAccount);
-
-        // Установка дат
-        financialGoal.setStartDate(startDate == null || startDate.isEmpty() ? LocalDate.now() : LocalDate.parse(startDate));
-        if (endDate != null && !endDate.isEmpty()) {
-            financialGoal.setEndDate(LocalDate.parse(endDate));
-        }
-
-        GoalStatus goalStatus = GoalStatus.valueOf(status.toUpperCase());
-        financialGoal.setStatus(goalStatus);
-
-        financialGoalService.saveFinancialGoal(financialGoal);
-        return "redirect:/financial-goals";
+        financialGoalService.createFinancialGoal(bankAccountId, name, targetAmount, startDate, endDate, status);
+        return "redirect:/financialGoals/financial-goals";
     }
 
 
@@ -107,14 +78,15 @@ public class FinancialGoalPageController {
         model.addAttribute("accumulated", accumulated);
         model.addAttribute("progressPercentage", progressPercentage);
 
-        return "goal-transactions";
+        return "transactions/goal-transactions";
     }
+
     @GetMapping("/edit/{id}")
     public String editFinancialGoalForm(@PathVariable Long id, Model model) {
         FinancialGoal financialGoal = financialGoalService.getFinancialGoalById(id)
                 .orElseThrow(() -> new RuntimeException("Financial Goal not found"));
         model.addAttribute("financialGoal", financialGoal);
-        return "edit-financial-goal";
+        return "financialGoals/edit-financial-goal";
     }
 
     @PostMapping("/edit/{id}")
@@ -125,29 +97,14 @@ public class FinancialGoalPageController {
                                           @RequestParam String endDate,
                                           @RequestParam String status) {
 
-//        FinancialGoal financialGoal = financialGoalService.getFinancialGoalById(id)
-//                .orElseThrow(() -> new RuntimeException("Financial Goal not found"));
-        FinancialGoal financialGoal = financialGoalRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Financial Goal not found"));
-        financialGoal.setName(name);
-        financialGoal.setTargetAmount(targetAmount);
-        if(startDate != null && !startDate.isEmpty()) {
-            financialGoal.setStartDate(LocalDate.parse(startDate));
-        }
-        if(endDate != null && !endDate.isEmpty()) {
-            financialGoal.setEndDate(LocalDate.parse(endDate));
-
-        }
-        GoalStatus goalStatus = GoalStatus.valueOf(status.toUpperCase());
-        financialGoal.setStatus(goalStatus);
-        financialGoalRepository.save(financialGoal);
-        return "redirect:/financial-goals";
+        financialGoalService.updateFinancialGoalForm(id, name, targetAmount, startDate, endDate, status);
+        return "redirect:/financialGoals/financial-goals";
     }
 
     @DeleteMapping("/delete/{id}")
     public String deleteFinancialGoal(@PathVariable Long id) {
         financialGoalService.deleteFinancialGoal(id);
-        return "redirect:/financial-goals";
+        return "redirect:/financialGoals/financial-goals";
     }
 
     @GetMapping("/{goalId}/create-transaction")
@@ -157,7 +114,7 @@ public class FinancialGoalPageController {
 
         model.addAttribute("financialGoal", financialGoal);
         model.addAttribute("transaction", new Transaction());
-        return "create-financial-goal-transaction";
+        return "transactions/create-financial-goal-transaction";
     }
 
 
@@ -169,36 +126,8 @@ public class FinancialGoalPageController {
             @RequestParam(required = false) String transactionDate,
             @RequestParam String type) {
 
-        FinancialGoal financialGoal = financialGoalService.getFinancialGoalById(goalId)
-                .orElseThrow(() -> new RuntimeException("Financial Goal not found"));
-
-        Transaction transaction = new Transaction();
-        transaction.setName(name);
-        transaction.setAmount(amount);
-        transaction.setType(TransactionType.valueOf(type.toUpperCase()));
-        transaction.setFinancialGoal(financialGoal);
-
-        LocalDate date = (transactionDate == null || transactionDate.isEmpty())
-                ? LocalDate.now()
-                : LocalDate.parse(transactionDate);
-        transaction.setTransactionDate(date);
-
-        BankAccount bankAccount = financialGoal.getBankAccount();
-
-        // Изменяем баланс счёта на основе типа транзакции
-        if (transaction.getType() == TransactionType.EXPENSE) { // исходящая - накопления из копилки достаются
-            bankAccount.setBalance(bankAccount.getBalance().add(amount));
-        } else if (transaction.getType() == TransactionType.INCOME) { // входящая для фин цели - значит со своего счета перечисляем в копилку
-            if (bankAccount.getBalance().compareTo(amount) < 0) {
-                throw new IllegalArgumentException("Insufficient funds for this transaction.");
-            }
-            bankAccount.setBalance(bankAccount.getBalance().subtract(amount));
-        }
-
-        transactionService.saveTransaction(transaction);
-        bankAccountService.updateBankAccount(bankAccount);
-
-        return "redirect:/financial-goals/" + goalId + "/transactions";
+        financialGoalService.createTransactionForGoal(goalId, name, amount, transactionDate, type);
+        return "redirect:/financialGoals/financial-goals/" + goalId + "/transactions/transactions";
     }
 
 }
