+++ src/main/java/ru/froleod/FinanceTrackerApp/controller/LogoutController.java
@@ -8,6 +8,6 @@ public class LogoutController {
 
     @GetMapping("/logout")
     public String logout() {
-        return "redirect:/login?logout";
+        return "redirect:/auth/login?logout";
     }
 }
