+++ src/main/java/ru/froleod/FinanceTrackerApp/repo/BankAccountRepository.java
@@ -7,7 +7,5 @@ import ru.froleod.FinanceTrackerApp.model.BankAccount;
 @Repository
 public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
 
-    BankAccount findByUserId(Long userId);
-
     BankAccount findBankAccountByUserUsername(String username);
 }
