+++ src/main/java/ru/froleod/FinanceTrackerApp/repo/BudgetRepository.java
@@ -11,8 +11,6 @@ public interface BudgetRepository extends JpaRepository<Budget, Long> {
 
     List<Budget> findByBankAccountId(Long accountId);
 
-    List<Budget> findByBankAccountUserId(Long userId);
-
     List<Budget> findByBankAccountUserUsername(String username);
 
 }
