+++ src/main/java/ru/froleod/FinanceTrackerApp/repo/FinancialGoalRepository.java
@@ -4,10 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.froleod.FinanceTrackerApp.model.FinancialGoal;
 
-import java.util.List;
-
 @Repository
 public interface FinancialGoalRepository extends JpaRepository<FinancialGoal, Long> {
 
-    List<FinancialGoal> findByBankAccountUserId(long userId);
 }
