+++ src/main/resources/templates/auth/registration.html
