+++ src/main/resources/templates/transactions/create-financial-goal-transaction.html
