+++ src/main/resources/templates/transactions/create-transaction.html
